job analysis function added

+++ src/main/java/group/service/iko/dto/HistoryRecordDTO.java
@@ -13,34 +13,34 @@ import java.util.List;
 
 public class HistoryRecordDTO {
 
-    private  int id;
-    private  String title;
+    private int id;
+    private String title;
     private String recordDate;
     private int SMR;
     private String recordInformation;
     private String machine;
     private String otherInfo;
-    private int laborHour;
+    private double laborHour;
 
     public HistoryRecordDTO(HistoryRecord historyRecord) {
-        id =historyRecord.getId();
+        id = historyRecord.getId();
         title = historyRecord.getTitle();
         recordDate = CalendarAdapter.getStringFormat(historyRecord.getRecordDate());
         SMR = historyRecord.getSMR();
         recordInformation = historyRecord.getRecordInformation();
         Machine machineOfRecord = historyRecord.getMachine();
-        machine = machineOfRecord.getModel() +"; sn " +machineOfRecord.getSerialNumber();
+        machine = machineOfRecord.getModel() + "; sn " + machineOfRecord.getSerialNumber();
         otherInfo = historyRecord.getOtherInfo();
         laborHour = historyRecord.getLaborHour();
-            }
+    }
 
-            public static List<HistoryRecordDTO> transformIntoDTO(List<HistoryRecord> recordList){
-        List<HistoryRecordDTO > recordDTOList = new ArrayList<HistoryRecordDTO>();
-        for(HistoryRecord historyRecord : recordList){
+    public static List<HistoryRecordDTO> transformIntoDTO(List<HistoryRecord> recordList) {
+        List<HistoryRecordDTO> recordDTOList = new ArrayList<HistoryRecordDTO>();
+        for (HistoryRecord historyRecord : recordList) {
             recordDTOList.add(new HistoryRecordDTO(historyRecord));
         }
         return recordDTOList;
-            }
+    }
 
 
     @Override
@@ -113,11 +113,11 @@ public class HistoryRecordDTO {
         this.otherInfo = otherInfo;
     }
 
-    public int getLaborHour() {
+    public double getLaborHour() {
         return laborHour;
     }
 
-    public void setLaborHour(int laborHour) {
+    public void setLaborHour(double laborHour) {
         this.laborHour = laborHour;
     }
 }
